Kindof fix to bug that Moveit doesn't add collision object
This seems extremely weird:
If we call add_ground_plane() in the GraspPlanner constructor, the object
is not added to the scene (this can be seen in Rviz).
However, if we add add_ground_plane() in the plan_arm_pose() method, then
it suddenly works which doesn't make sense at all!
Seems to be a weird Moveit bug...